changing from CityGMLReader to FeatureCityGMLReader
